requirements_server.txt
# SMS Database API Server Requirements

# Core Flask API
flask>=2.0.0

# Environment configuration
python-dotenv>=0.19.0

# Hardware GPIO control
pigpio>=1.78

# Serial communication
pyserial>=3.4

# HTTP requests
requests>=2.25.0

# OLED Display (lightweight, no PIL required)
luma.oled>=3.8.0

# Thread and async support
threading-timer>=0.1.0

# OLED Display (SSD1306)
luma.oled>=3.8.0
pillow>=8.0.0

# Standard libraries (usually included with Python)
# sqlite3 - built-in
# threading - built-in
# queue - built-in
# subprocess - built-in
# datetime - built-in
# time - built-in
# re - built-in
# os - built-in
